backend/requirements.txt
flask
flask-cors
firebase-admin
requests
python-dotenv
flask-mail 